fix postprocess burn issue

+++ cosmosis/postprocessing/postprocess.py
@@ -104,7 +104,7 @@ class EmceeProcessor(MetropolisHastingsProcessor):
 		thin = self.options.get("thin", 1)
 
 		if 0.0<burn<1.0:
-			burn = len(cols[0])*burn
+			burn = int(len(cols[0])*burn)
 		else:
 			burn = int(burn)
